Week 6 - Game Art Part 3/Importing to Unity & Building Scenes
Today we imported all our Maya Assets and the textures we made in Substance Painter into Unity, learning how to apply the texture maps onto materials and applying them on the LP Models as well as how to properly position the models for our level. We also briefly went through different lighting and how to apply a skybox texture.

Assessment Task 1 is due this Sunday, so this game needs to be polished up and built for submission. Brendan's posted a summary on Teams but essentially the game needs to:
1. Control properly with a simple complete level with a win condition.
2. Have functioning UI (Preferably custom and with more stuff than what we did in class like implementing bars, map, another menu etc.)
3. Complete Modular Assets with textures made in Substance Painter 3D.

For me, create more interesting UI and finish building the level using the modular assets. And if you have time, find a way to control the camera with the mouse in-game and create a BGM. You got this.