Quick fix
I think I fixed the problem. It may have something to do with the pickup count not resetting when the level restarts, so the currentPickUpCount == maxPickUpCount when the scene restarts, causing the game to automatically win. I don't think that was the actual cause because the console kept saying it was due to the timer but, eh, it works. Back to finetuning before submission.