Game technically ready for submission, however...
I didn't need to fine-tune much, game's practically ready for submission. But I REALLY want to make a BGM for this game, like a short looping ambient track. Whether I do or don't add it will depend on time or willingness on my part and I am definitely willing to.