Added a Free Camera mode... kinda?
The camera now follows the position of the mouse and rotates along a pivot connected to the player. The camera's no longer in a fixed spot, but I didn't get the results I wanted. The player acts independently from the camera and moves along the same axis indifferent to the camera's position in the world. I also only wanted to the control the camera when the mouse button is held down, just like in the editor, and not just the mouse it itself controlling the camera. I also want to be able to zoom in and out. But HEY, it's just the first step. I also modified the level so it's better suited for tests. I intend to use it in a series of tutorial levels as well.