Steps back
Writing this so I don't back myself into a rut. I need to remind myself what I want my game to achieve in the end. My version of Roll-A-Ball wants to be open and invites exploring. The level design is the most important aspect of this game, so let's plan and design these ahead of time and worry about other assets later. We'll get the setting just right to tell your story.