Week 9 - Laying out the first level.
I've decided on four levels to base my game around, the tutorial landfill, the expansive streets, the momentum based underground, the horizontal rising Office building, and the hazardous lab. As of now, I've laid out the first level and what I want from it. It's the first level so it's pretty linear while introducing the controls and potential hazards. I also want to do something thematic with the landfill, where the player returns to the landfill to evoke a full circle feeling. I'll have to figure out the game's story for that, after laying out the rest of the levels and mechanics.