Starting to stress out...
I brought this upon myself, laying it off so long. There's still so much to do and ideas I want to implement but I can't... bring myself to act on it. I told myself that that would change and it would be brighter and hopeful yet...

... Finish at least laying out the second level. I'm not sure how far I can go with the assets and cutscenes I planned, and there's still other assessments that are soon due or overdue.

Typing these all out helps. It doesn't change my current situation but... it helps to compose yourself.

... take a break for now.